+++ plugins/eg-scope.lv2/examploscope.c
@@ -249,7 +249,7 @@ run(LV2_Handle handle, uint32_t n_samples)
 
     // Add UI state as properties
     lv2_atom_forge_key(&self->forge, self->uris.ui_spp);
-    lv2_atom_forge_int(&self->forge, self->ui_spp);
+    lv2_atom_forge_int(&self->forge, (int32_t)self->ui_spp);
     lv2_atom_forge_key(&self->forge, self->uris.ui_amp);
     lv2_atom_forge_float(&self->forge, self->ui_amp);
     lv2_atom_forge_key(&self->forge, self->uris.param_sampleRate);
@@ -295,7 +295,8 @@ run(LV2_Handle handle, uint32_t n_samples)
   for (uint32_t c = 0; c < self->n_channels; ++c) {
     if (self->ui_active) {
       // If UI is active, send raw audio data to UI
-      tx_rawaudio(&self->forge, &self->uris, c, n_samples, self->input[c]);
+      tx_rawaudio(
+        &self->forge, &self->uris, (int32_t)c, n_samples, self->input[c]);
     }
     // If not processing audio in-place, forward audio
     if (self->input[c] != self->output[c]) {
